fixed a lot of stuff
