+++ src/components/styles.css
@@ -20,4 +20,10 @@ body {
    max-width: 500px;
    max-height: 500px;
    object-fit: cover;
+   border-radius: 15px;
+}
+
+.newRecipe {
+   border: 2px dashed black;
+   height: 100%;
 }
